Refactored TabbedInventoryActivity UI Files
